FessScrapper_project/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==3.2.18
djangorestframework==3.15.1
djongo>=1.3.6
dnspython==2.6.1
gunicorn==22.0.0
idna==3.7
pymongo==4.7.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
soupsieve==2.5
sqlparse>=0.2.4
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1